+++ src/utils/toKebabCase.js
@@ -1,2 +1,0 @@
-export const toKebabCase = (str) =>
-  str?.replace(/([a-z])([A-Z])/g, '$1-$2').replace(/\s+/g, '-').toLowerCase()